requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.0
djmail==2.0.0
drf-spectacular==0.24.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.16.0
lxml==4.9.1
model-bakery==1.7.0
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
premailer==3.10.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-cryptography==3.3.23
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
